Updated openstack/openstack Project: openstack/tripleo-heat-templates  2399cb85144008bd1a2705809ccef2109a453905
Fix dependencies for PreConfig/PostConfig resources

We have a circular dependency errror since

https://review.openstack.org/#/c/452734/ landed.

This adjusts the dependencies to ensure we run pre-config before
the first puppet deploy step, and removes the duplicate declaration
of the ControllerPostConfig resource.  Also we ensure the first
container step always depends on the same step puppet deploy.

Change-Id: I70c5a39fb36b951bdeb04c15bddac7d00eebf08a
Closes-Bug: #1686098