Updated openstack/openstack Project: openstack/nova  671eb66f3defb66b774b5e34020add43841d33a3
Clean up ClientRouter debt

This removes the two deprecated by_instance() and by_host() methods
in ClientRouter and replaces them with calls to a single client()
routine. We stopped using the instance and host parameters in these
methods in an earlier patch to remove the caching in the router.

Related to blueprint cells-aware-api

Change-Id: Ifa7cd99a442ef0f8e9ad4a6b52982d0fedf1cd7e