Updated openstack/openstack Project: openstack-infra/system-config  d3ebb0bde3aca01471d3381221c4ca36646c07be
Add mosquitto_tls_ca_file to public hiera

The mosquitto_tls_ca_file is just the contents of the ca bundle from
instantssl. There is nothing secret about it and you can just download
a copy of it from comodo/instantssl. This patch just moves it to the
public hiera file.

Change-Id: I2438ed14887962a605a534aba60d58e7b0fee23a