Updated openstack/openstack Project: openstack/training-guides  9d28a6ead4778180e28a331e0fda99d4ede3deea
[upstream] fix links and wording in Sandbox slides

A number of the links from these slides were no longer accurate.
This patch set fixes those links.  Note that link for the commit
message slide is set based on the new patchset that is currently
under review for those slides.  It also updates the wording on
a number of the slides for clarity.

A note is also added that they may need to use 'git-review' instead
of 'git review' depending on the environment they are doing
development in.

Change-Id: Ibd7405f062ab2b119fba21ebb7f8138455f7fab3