Updated openstack/openstack Project: openstack/training-guides  cf102483e0e9b2b66b49085dce05965229c48e15
[upstream] Update release-cycle slides

This patch pushes up changes to the howitsmade-release-cycle.rst
slides.  The goal of the patch is clean-up wording on a number of
slides, tie in the idea of the Project Teams Gathering and add
more exercises to get the class looking up the information on their
own.  Also good opportunities for the class to talk in their groups.

I have also improved notes on a number of slides to provide more
guidance to the presenter and consolidated some of the wordier slides.

Change-Id: I7632c2e855844e6d8aa0b7c75734c6be3f5ae9e1