Updated openstack/openstack Project: openstack/tempest  a55be93c288c86fe832e197332533b6ae262e11a
Remove skip condition when resize is enable

This commit removes the skip condition with flavor_ref and
flavor_ref_alt when the resize feature is enable. Because when
resize feature is enable, the flavor_ref and flavor_ref_alt should be
set properly, otherwise, it is a misconfiguration. So we should fail
but not skip at the situation.

And This commit also add a help message for the resize config to avoid
users confusing.

Change-Id: I8eab2e57eb77e7a1a8dae7ffd63ec95ae6a026c3