Updated openstack/openstack Project: openstack/training-guides  3e7444509273b5aa0909a2e5f551437c66dc70ff
[upstream] Add steps to checkout/configure/run DevStack

The Demo Environment slides have an exercise to make sure that
DevStack is checked out to the VM where they wish to run DevStack
but have no instructions for doing so.  I added a slide with the
steps for cloning, configuring and running devstack.

I also improved the wording of the exercise on the final slide.
The details on the slide for using 'screen' were pretty light so
I also added content there.

Change-Id: I96a623323020a630d88d5f6d8659a6d49bd44ec7