Updated openstack/openstack Project: openstack/ironic  075e80979156e7afeb603b6a72a1bd6b1943dd9e
VirtualPDU use libvirt group instead of libvirtd

Currently, VirtualPDU uses 'libvirtd' group, which errors since this
group is not used anymore on ubuntu because we now use the same group
that debian uses 'libvirt'.

This commit change the used group to 'libvirt'. This is the right group
and it's allowed to do changes on virsh.

I would rather use LIBVIRT_GROUP however a bug in the selection logic
sets it to 'libvirtd' group as soon as it exists, even if both exists.

Change-Id: Iddb61356b8a92184ca60a078ea9d929e5b51e629