Updated openstack/openstack Project: openstack/ironic  6f1e8db944a8207ce437bbf538081564f223372a
Read disk identifier after config drive setup

If original disk image has ID of 0x00000000 it's overwritten by parted
with a non-zero value. The original ID is stored in pxelinux.cfg/* config
file and the node fails to boot.

Reading the ID was moved to after config drive setup so that the final ID
is returned.

Change-Id: Icbfa3e93bf6252177a0baa5f89bf4565ff331520
Closes-Bug: #1685093