Updated openstack/openstack Project: openstack/cinder  3458923aafd66bc7f3f717ccfd8b0a6bb1cf96c2
GPFS: Enhance GPFS driver to work in container

This patch enhances the GPFS driver to support the
use case when cinder service is running inside Linux
container hosted on Nova-Compute node (This kind of an
environment is deployed by OpenStack Juju Charm).
With this change, driver would talk to GPFS client
on Nova-Compute nodes over ssh.

Change-Id: Ib3235999d1b34502cdcc8efc359cfc6f5df0412a