Updated openstack/openstack Project: openstack/cinder  c3d293dcd98d9bd4604dae6768dc956acc0bd13e
VNX: add missing group snapshot pool stats

Create group/volume failed when 'consistent_group_snapshot_enabled'
is set to '<is> True' in the group type specs and volume type extra
specs. This is because 'consistent_group_snapshot_enabled = True'
is in the stats for the backend but it is not in the stats at the
pool level.

In this patch, 'consistent_group_snapshot_enabled = True' is added
to the pool stats reporting to fix this problem.

Note: This patch should be backported to Ocata.
'consistencygroup_support=True' and consistencygroup driver
interfaces are not removed in this patch because this patch should
be backported. There will be a followup patch to remove
consistencygroup driver interface implementation from this driver.

Change-Id: I330fc93247a2509d961acfda978c3881391e01f2
Closes-Bug: #1681305