Updated openstack/openstack Project: openstack/trove  b0a45c0962317dfd6ca71898867cb75ee2aa0039
Fixed MySQL 5.7 not starting after restore

Resetting the root password in the guestagent is incompatible with
changes made in MySQL 5.7 (namely the removal of the password column).
The password is now reset using 'SET PASSWORD' as indicated in
http://dev.mysql.com/doc/refman/5.6/en/resetting-permissions.html
(this works with 5.7 as well). As of 5.7.6 the 'ALTER USER' command
may be used as well, however the issue of having restore strategies
for subversions of a manager would need to be addressed first.

Change-Id: I9453f3cdc6338942da4585a13a1d07a235604202
Closes-Bug: 1655715