Updated openstack/openstack Project: openstack/python-openstackclient  a28364a37a3eb76b41e4ef5d53f21e58d2094f0d
Fix NoneType error for volume snapshot create command

In volume snapshot command, <volume> is the same
as <snapshot-name> when --volume is not specified,
but <volume> cannot be None, so when <snapshot-name>
is not specified (<snapshot-name> is None), a NoneType
error appears.
So make <snapshot-name> no longer optional, it should
be always present.

Change-Id: I3d9f10753a8ef601e70816421c160598e2cc811f
Closes-bug: #1659894