Updated openstack/openstack Project: openstack/python-openstackclient  0a0bcbbda2c7384f6308bb8041119d6c38915e2f
Remove ipdb installation in tox debug section

The lastest ipdb depends on ipython 6.0, but ipython 6.0
only can be installed in python 3.3 and above, see
http://paste.openstack.org/show/607632/ . If we try to
run "tox -e debug" in python2.7, the install error is raised
and block the function. Remove the ipdb installation, it's
not necessary, we can use pdb to replace.

Change-Id: Ib47bb5925b7a5b1d3a319b58fa219c1b57dccb93