Updated openstack/openstack Project: openstack-infra/elastic-recheck  797f6ce2774fcff96ef87831c4ab7b14727911db
Track job timeouts in e-r

So that timed out jobs are classified as timing out and not noise in the
uncategorized list lets track these generically. We can also have
specific timeout bugs tracked in addition to this one (and at least one
is the the neutron multinode linuxbridge test).

Change-Id: I18b6f312ea3e81755c23285287c4192d8dd87a14
Related-bug: 1686542