Updated openstack/openstack Project: openstack/sahara-image-elements  ffb7f6889533a1e2a1a27a1838e7f753bf293f09
Fix: set the Fedora-specific ssh_config file for augeas

The Host entry in /etc/ssh/ssh_config file on Fedora 25+ is commented
because there is an Include directive and the custom settings have
been moved to a file there.
This leads to a failure when the augeas lens tries to write to that
files. If I read the schema correctly, the lens assumes that the Host
entry is available/uncommented (this it was the case before in Fedora,
and in CentOS and RHEL, and in Debian/Ubuntu, where a customization
is a applied to the upstream default file, which is used as it is now
in Fedora).
This maybe could be an issue of the lens which may not assume that
key, but I'm not sure whether is possible.
In the meantime, use the existing customization file on Fedora,
instead of /etc/ssh/ssh_config.

Change-Id: Iee5c109b1265b54a75f675d26c9fec49c777d298