Updated openstack/openstack Project: openstack-dev/grenade  6eeea326d99e3804c9add2f343cf846ae887b266
Set NOVA_USE_MOD_WSGI False for target

This ensure that for Pike we will not use the new uwsgi
functionality for the nova-api anod nova-metadata services, ensuring
that we continue to use the default infrastructure over the
release transition. This can be removed when both cycles are using
uwsgi.

Change-Id: I7b22692ea9135547d93b08724f1f9ef060f13007