Updated openstack/openstack Project: openstack/ceilometer  c956613b5a9aa22d45c4b570028d6302925d4b5c
Format the abstraction layer of inspector methods to avoid TypeError

There are several inspector methods definiton of the inspector base class
missed a "duration" parameter, which will raise an TypeError every polling
cycle if the methods didn't implemented in real inspector classes.
Normally, the not implemented inspector methods should raise a
ceilometer.NotImplementedError error which will be catched and permanently
skip the pollsters.

Change-Id: I1db6b2e59c2b359b27d86531d1ba0b41146bb69c
Closes-Bug: #1680657