Updated openstack/openstack Project: openstack-infra/puppet-subunit2sql  35025f8df5576718389f9d7da3e8865acc70710a
Move paho-mqtt package install to worker node

The paho-mqtt package is only ever needed on the worker because that is
what emits mqtt. However by putting it in init this also gets installed
on the server node. Additionally this commit adds a duplication check
around the package because it's possible that something else in another
module will conflict with paho-mqtt as we start adding mqtt to more
services.

Change-Id: Id0308d2d4d1843fcca73f459cffa2ae944bebd0c