Updated openstack/openstack Project: openstack-infra/project-config  bf77be31057e7be8bb4f432b1bdcb5e71192b0ff
Create experimental job for dox to test docker images

We eventually want to publish docker images to hub.docker.com.  Today,
we can experiment with dox, since we already have some images
published.   Once we establish how this work for dox, we can start the
process of rolling this out to other projects.

Change-Id: I048b923c0d8918e0bfcdf7030b4530251846d565
Signed-off-by: Paul Belanger <pabelanger@redhat.com>