Updated openstack/openstack Project: openstack-infra/project-config  16138f5de0714de1586f32bfdbf52525c069272c
Fix metric names for received messages

This commit fixes a "typo" in the stat names for the metrics with
messages received. In statsd the stats are spelled incorrectly which
means these didn't match anything. Correcting this to match my
inability to spell in the mqtt_statsd config and we should have working
graphs.

Change-Id: Ia0965f12239190c6dd11a6a0d78bf8f18fe178c9