Updated openstack/openstack Project: openstack-infra/system-config  09ff9ffff16456b2fb1e9351fd3b1077fe78c6a5
Remove references to the simple_slave class

This class is not used anywhere, which we know because it includes the
openstack_project::base class which doesn't exist. We also remove,
rather than update, the documentation's reference to it because pretty
soon devstack-gate won't need the openstack_project module at all.

Change-Id: I44ac578af7c659adef614f73b250523d1acf09c4