Updated openstack/openstack Project: openstack-infra/project-config  7d49a1f1f1915c130759a1967320b8f9e21666eb
OSA: Enable cross-repo testing for integration testing

The ZUUL_* environment variables are not set when the job
is executed using sudo. Instead of using 'sudo -E' which
would pass along a whole bunch of env vars which may
interfere with the build, we only do a minimal set in
order to enable cross-repository testing.

Change-Id: If9c8bcb0069d6d7e45ac2a484f3bb97dbf80772b