Updated openstack/openstack Project: openstack/training-guides  1d7cd3f696ed28386e72b67b439b3512d9c3b53f
[upstream] Improve wording in the Closing Exercise slides

Some of the steps on the last page of the Closing Exercise slides
was not clear.  This patch rewords a couple of the bullet points
to make it clearer what the students should do.

This patch also moves this content to the archive section given that
we did not feel we would get to this during the training session.
The team feels that time is better spent on individual project
deep dives.

Change-Id: I2da7be6fd7a589838b9fb71aebab1d60a998dc70