Updated openstack/openstack Project: openstack-infra/puppet-logstash  2f689ee89968591cc41d0510b7e33daade2cc8c0
Remove logstash mqtt plugin install

The logstash mqtt plugin caused issues in production and we're not going
to use it. We'll be emitting mqtt notifications directly from the
gearman worker instead. So this commit removes the plugin install so we
can set enable_mqtt to true to install the cert.

Change-Id: I0a17444cce18dd4b63f1f924e393483f6d8fe8eb