Updated openstack/openstack Project: openstack-infra/tripleo-ci  c6c3a211ca07cb6fa4b570ead06a366bdbf1e5e2
Support container upgrades in multinode OOOQ CI

Add upgrade to the multinode.yml playbook. The current use is for
upgrade from non-containerized to containerized deployment. The addition
could be used for entirely non-containerized upgrade as well, provided
that correct OOOQ config is used, but we may never get to utilize this.

Change-Id: I4c7599d56a5a9cfed652c246725bc986aba7d3ea
Depends-On: Ibc71da44335c50ae615c331aed6bb4d2d1203649
Depends-On: Ia013fe331d84adce7d36af6bd4e33d37c0755fdd