Updated openstack/openstack Project: openstack/nova  26117cf544be6b2f907200a474c3cffd0ff1d0f3
Remove useless API tests

The 'test_flavor_swap' tests are validating behavior that's already
validated by the larger 'test_flavors' tests and a plethora of
functional tests. This is legacy from the time where exposing 'swap' was
required the 'os-flavor-swap' extension [1]. This extension was not
ported to v2.1 [2] and 'swap' is now part of the default response.

Stop special casing these particular fields.

[1] https://github.com/openstack/nova/commit/81dd5b78
[2] https://github.com/openstack/nova/commit/a31d917a

Change-Id: I55a57b76a92677fbe7db9b6ef91f8cd4daf30d87