Updated openstack/openstack Project: openstack-infra/project-config  e41d2cb1fe864582f2db792101f78d9f30f0be77
OSA: Output ZUUL variables to a file

The current mechanism implemented in
https://review.openstack.org/458890 does not work
and causes jobs to fail when multiple dependent
patches are included in a patch. This is due to
the spacing in the values of ZUUL_CHANGE_IDS.

This mechanism instead outputs the data to a
file which can be appropriately sourced by the
functional test script to achieve the same goal
and be less of a maintenance burden.

Change-Id: If9ff68eec9aa900680ed846f72d182e4115658ad