Updated openstack/openstack Project: openstack-infra/project-config  5279a088744764d73be98819297ce05b944865fa
Reproduce non-puppet parts from install_puppet.sh

We want to stop installing puppet on nodepool nodes, but the
install_puppet.sh script does more than just that. This patch adds some
of the packages that install_puppet.sh was installing that are unrelated
to puppet or needed for more than just puppet.

Change-Id: Ia9c45b762eea9f2989c9c63fb944b9e5e1f17ed2