Updated openstack/openstack Project: openstack/keystonemiddleware  08bf885cebf3286bc297b8d7948dc569bf755ff9
Bump the token deferral message from info to debug

In Swift, we have a public access endpoint "/info", which
returns basics like version number, maximum object name length,
and such. When a user accesses it, the request travels the
pipeline, including authtoken, and since it has no valid token,
a message is logged. At some installations, monitoring agents
poll "/info" to see if Swift is up, and this floods the logs.

It would be best for us to dispose with this message altogether,
but I am concerned that operators of other services may find
this message useful. So, let's only mark it as a debug level.

Change-Id: I77abc3809a91e381b7650a9955046fe6d72a8089