Updated openstack/openstack Project: openstack-infra/puppet-subunit2sql  c2beb421055e18234723b35f9931457200f680d3
Add source_url to mqtt payload

This commit adds the source_url to the mqtt payload for events emitted
by the subunit worker. Just the build_uuid isn't enough to figure out
what's going on because we typically do 2 events for each uuid, one with
a .gz extension and the other not. We need to distinguish between the 2
to figure out what's actually going on.

Change-Id: I56f0b9fb1128b412cd88e49164eb91c3e9e3e4cb