Updated openstack/openstack Project: openstack/barbican  7f9839707a4a6c011a86cf3c4e12ca5d9380491f
Fixes for Running Functional Tests with Python 3.5

Fixing the standard things when making code py27/p35 compatible.

Also, removing the logging of the passed value of an HTTP
header.  If the value could not be encoded to log, then there
are Tracebacks that showed up with Python 3.5.  Since the value
can be passed by a user, it should either be scrubbed before logging
or not logged, to prevent possible content injection in the log
stream.

Change-Id: I8df1553acb6c7e5f75a1b50f024dc032ca982a93