Updated openstack/openstack Project: openstack-infra/tripleo-ci  48aaaace125cbc71e7075e4b8741e656853802df
Redirect stack-show output in tripleo.sh

We use these functions in CI, and it bloats the console.log because
we dump the entire stack twice, and it's already stored in the postci
log.  We only care if the stack-show returns zero (e.g the stack exists)
so redirect any output.

Change-Id: I70f9b47606d3b3cf7683bcde379c0609ff48a72d