Updated openstack/openstack Project: openstack-infra/project-config  b18e11c0f20deaad34bba84ada1ed288e7a5b190
Let the jenkins user own the cache files

Earlier we placed these files in the jenkins user's homedir. That was
causing a problem so the cache directory was moved into /opt and
symlinked for backwards compatibility[1]. However, the jenkins user
doesn't have write access to directories created by root and will fail
to write files to that directory in devstack-gate[2][3]. This patch
changes the ownership of the cache directory back to the jenkins user so
that it should work the same as it worked when the directory was in the
jenkins user's homedir.

[1] https://review.openstack.org/#/c/460835/
[2] http://git.openstack.org/cgit/openstack-infra/devstack-gate/tree/functions.sh#n516
[3] http://paste.openstack.org/show/608377

Change-Id: I275f6ac8e5bb71aef904e8b9406a8810dae9bc67