Updated openstack/openstack Project: openstack/nova  98f2071e06efd928530f0347535a2f344e4131ec
Add reserved_host_cpus option

When discussing for the blueprint placement-claims we had a consensus on
using reserved values for helping operators to provide overhead limits
for all the instances.

Since we only had 2 opts for RAM and disk, we missed a possible CPU one.
This patch is adding that one.

Change-Id: Ic3162f0ab39a703798b4a2a7e860fed628da1e7e
Partially-Implements: blueprint placement-claims