Updated openstack/openstack Project: openstack-infra/project-config  217a956a1953500f6ade3b68fac19c4502a2c184
Add messages dropped to mosquitto graphs and add units

This commit cleans up the grafana dashboard for mosquitto a bit by
adding units to the titles so we know what the data we're looking at is.
Additionally it adds a new line for messages dropped which is another
stat we're populating from mqtt_statsd.

Change-Id: I4940c0fbf48ddd41c745d643bc95652bb5eaa9d4