Updated openstack/openstack Project: openstack/training-guides  3d8542b3753cde679c83a14241ab5c01a79683ee
[upstream] Generalize Jenkins slides to Zuul

During review of the upstream training slides is was decided
that the existing detailed overview of Jenkins was not as
helpful.  We decided to generalize the topic a bit to talk
about Zuul and the different types of gates.

This patch removes the old content and renames the slides
accordingly.  The content now introduces status.openstack.org ,
explains Zuul and the different pipelines.  Talks about how to
find logs and investigate elastic-recheck failures.  As with
other updates things have are more exercise focused.

Change-Id: If527a57ee3874aa0168d08f093937970f90b1d10