Updated openstack/openstack Project: openstack-infra/shade  59d37e89a1431195f1fbb8f8ab80f9e59239c56c
Remove two lines that are leftover and broken

Not really sure how these crept in - they never would have made sense.
Also not sure how this codepath _worked_ in the live test I used (with
the results in IRC) - but they definitely break as they should when
triggered in unit tests.

Change-Id: Iddc69cbd325c7a3a25cde48a2b2bddb01826c0ff