Updated openstack/openstack Project: openstack-infra/project-config  a71731e19bfd53d6e7058f8217a09cb6143e4a91
TripleO: experimental container upgrade job on more repos

So far we've only been running the container upgrade job on the
tripleo-ci repository. The job is now implemented and we got successful
runs with it recently, we should run it on repos we run the normal
upgrade job. Experimental and non-voting for now.

Change-Id: I410acaed5a263b529a8c56badea52561096b5ec3