Updated openstack/openstack Project: openstack-infra/project-config  789bbb1bf8f90c0a50503a8149929885fb54cd77
neutron: add ovs firewall tempest job to check queue

The job has been in experimental for a while and is currently passing.
It was triggered only 3 times in the last month and always passed. This
patch requests to add the job as non-voting just to get better overview
of its stability. There are currenlty no known issues so we can make it
voting very soon if it's proven stable enough.

Change-Id: I7bf679e53a4c86a41fd12ce78a1361b62bbf36cc