Updated openstack/openstack Project: openstack-infra/project-config  6d0efa540eff0da102b7199a800a4dd058699533
Load plugin configuration for the ceph job

This will allow us to further refine the set of tests to run and we
maintain this configuration within the plugin repository.  The
configuration file path is wrapped in a conditional to prevent the job
from failing while the plugin configuration is being tested and refined.

Change-Id: Ib3ccd42720ee7f4464567737f0bc1a297c590b74