Updated openstack/openstack Project: openstack-infra/project-config  d69aaf4c4731c8d64e1d842c0c87993e376e7c88
Add experimental gate job for murano dashboard

The purpose for this experimental gate is to replace an outdated
third-party hosted ci environment that runs selenium tests for
murano dashboard. Currently, the third-party murano-ci is
causing continuous problems.

The solution is to create, for the time being, an experimental
gate for running selenium tests. Because these tests might take
longer to run on infra, only package tests will be executed
(because the gate is experimental, only a gate for package
tests will be added; based on performance and stability, either
all selenium tests will be run in one gate or more gates will be
added).

Change-Id: I908c24d3e732d42bb87144bd6bba8a8243f03be3