Updated openstack/openstack Project: openstack-infra/project-config  94c5d3a3c1a7744b6b2e113399bcc37804496233
Remove fedora and debian from puppet-apply jobs

The only puppet-apply test that was being run for these OS types was the
dummy single-use-slave-devstack node, which now does nothing. In order
to remove the dummy node from site.pp we need to stop trying to run
tests on these OS's, otherwise the apply test will fail due to being
unable to find a puppet node to run tests for.

Change-Id: Ib1c2f6085f9e8b40e6a78cfb898738b0950bc339