Updated openstack/openstack Project: openstack/cinder  c81e1e2713a7eeb74ab75061b54454a4ad1532e2
Give more time for LVM deactivation

Current code retries three times for deactivation to complete,
waiting one second between retries. On some heavily loaded
systems, and apparently the gate, this is not enough time for
the operation to complete.

This attempts to work around those slower systems by increasing
our retries to 5 and adding a backoff of 2 seconds to give it
more time.

Change-Id: I4f40a1984fe828c8ff965033f7e25b1d7516ab1e
Closes-bug: #1687044