Updated openstack/openstack Project: openstack-infra/infra-manual  a88d4831362dafae3ab26bfb5a9d2209491c91b5
Register on PyPI with a minimal PKG-INFO file

In recent months, pypi.python.org removed the detailed package
submission Web form. Now the next simplest option is to upload a
precreated PKG-INFO file, but at least a minimal one can be as few
as two lines (corresponding to the same two required fields in the
now-missing form).

Change-Id: Iadf0de6ae6105dff9d36472ddc7e0ddb648fbc97