Updated openstack/openstack Project: openstack/neutron  951e765bbfc68a845d58ed8ad23e538c897277b6
Only add "on-link" routes for L2 adjacent subnets

When multiple subnets exist on a single network, the
DHCP agent adds on-link routes for all of them since
they are in the same L2 network.  If either subnet
has a segment_id it can only be considered as on-link
if they match, else we should not include a subnet
route.

These extra routes are optional anyways according to
RFC 3442, but were added for the use case when all of
the subnets are considered adjacent, which allows
instances to bypass the router and communicate directly.

Closes-Bug: #1668145
Change-Id: Iae889e9226a61059cd4f3d37fbe48d013b7a3482
Implements: blueprint tripleo-routed-networks-deployment