Updated openstack/openstack Project: openstack/oslo.messaging  c8ee259bacf2b71fc717ef94f43df0f7d2c4cf2a
Retry support for oslo_messaging_notifications driver

When the underlying RPC is configured with rabbit along
with oslo_messaging_notifications, currently the retry is
not supported. There are retry support specific to drivers,
it's better to have the support uniform across the drivers,
by adding configuration retry on the oslo_messaging_notifier.

Closes-bug: 1666364

Change-Id: I8d547ef2a9e8b086f94a1efca17735f3f3cdaac1