Updated openstack/openstack Project: openstack/tempest  ee1108b15aebd15a3ab0aba7d2a28ad37ca709fe
Use min_count to create servers in ListServersNegativeTestJSON

This is to use min_count=3 to create 3 servers, and then delete
one of the created servers, so to reduce the time of waiting for
one server to be active.

Besides, there is only one deleted server so needn't use a list,
thus the code can be more simple.

Change-Id: I1daabb64147bb9e0025ae7d7a0913254c966704f